Refactors TeamStats
