+++ lib/stat_tracker.rb
@@ -1,27 +1,31 @@
-# require './lib/league_stats'
-# require './lib/team_stats'
-# require './lib/season_stats'
-# require './lib/list_mod'
-# require './lib/game_stats'
-# require './lib/season_most_least'
-# require './lib/avg_goals'
+require './lib/league_stats'
+require './lib/team_stats'
+require './lib/season_stats'
+require './lib/list_mod'
+require './lib/game_stats'
+require './lib/season_most_least'
+require './lib/avg_goals'
+require './lib/hash_module'
+require './lib/summary_module'
 
-require_relative './game'
-require_relative './team'
-require_relative './game_team'
-require_relative './game_repo'
-require_relative './team_repo'
-require_relative './game_team_repo'
-require_relative './league_stats'
-require_relative './team_stats'
-require_relative './season_stats'
-require_relative './list_mod'
-require_relative './season_most_least'
-require_relative './avg_goals'
-require_relative './game_stats'
+# require_relative './game'
+# require_relative './team'
+# require_relative './game_team'
+# require_relative './game_repo'
+# require_relative './team_repo'
+# require_relative './game_team_repo'
+# require_relative './league_stats'
+# require_relative './team_stats'
+# require_relative './season_stats'
+# require_relative './list_mod'
+# require_relative './season_most_least'
+# require_relative './avg_goals'
+# require_relative './game_stats'
+# require_relative './hash_module'
+# require_relative './summary_module'
 
-require "csv"
-require "pry"
+require 'csv'
+require 'pry'
 
 
 class StatTracker
@@ -32,13 +36,14 @@ class StatTracker
   include List
   include SeasonMostLeast
   include AvgGoals
+  include HashModule
+  include Summary
 
   attr_accessor :games, :game_teams, :teams
   def initialize
     @games = games
     @teams = teams
     @game_teams = game_teams
-
   end
 
   def self.from_csv(locations)
@@ -66,8 +71,6 @@ class StatTracker
       game.away_team_id
     end
   end
-
-
 
   def hash_games_by_season
     @games.repo.group_by do |game|
